redact-pdf/face-anonymisation/requirements.txt
fastapi
uvicorn
pymupdf
opencv-python-headless
numpy
pillow
python-multipart